+++ app/components/theme-select.tsx
@@ -1,5 +1,6 @@
 import { capitalize } from "@hiogawa/utils";
 import React from "react";
+import { useMatchMedia } from "../utils/hooks-client-utils";
 import { SelectWrapper } from "./misc";
 
 // by @hiogawa/utils-experimental/dist/theme-script.global.js
@@ -65,24 +66,3 @@ const FAVICON_LIGHT =
 
 const FAVICON_DARK =
   "data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24'%3E%3Cpath fill='none' d='M0 0h24v24H0z'%3E%3C/path%3E%3Cpath d='M18.5 10L22.9 21H20.745L19.544 18H15.454L14.255 21H12.101L16.5 10H18.5ZM10 2V4H16V6L14.0322 6.0006C13.2425 8.36616 11.9988 10.5057 10.4115 12.301C11.1344 12.9457 11.917 13.5176 12.7475 14.0079L11.9969 15.8855C10.9237 15.2781 9.91944 14.5524 8.99961 13.7249C7.21403 15.332 5.10914 16.5553 2.79891 17.2734L2.26257 15.3442C4.2385 14.7203 6.04543 13.6737 7.59042 12.3021C6.46277 11.0281 5.50873 9.57985 4.76742 8.00028L7.00684 8.00037C7.57018 9.03885 8.23979 10.0033 8.99967 10.877C10.2283 9.46508 11.2205 7.81616 11.9095 6.00101L2 6V4H8V2H10ZM17.5 12.8852L16.253 16H18.745L17.5 12.8852Z' fill='rgba(255,255,255,1)'%3E%3C/path%3E%3C/svg%3E";
-
-export function useMatchMedia(options: {
-  query: string;
-  defaultValue: boolean;
-}): boolean {
-  const result = React.useMemo(
-    () =>
-      typeof window === "undefined"
-        ? ({ matches: options.defaultValue } as MediaQueryList)
-        : window.matchMedia(options.query),
-    [options.query]
-  );
-  const rerender = React.useReducer((prev) => !prev, false)[1];
-
-  React.useEffect(() => {
-    result.addEventListener("change", rerender);
-    return () => result.removeEventListener("change", rerender);
-  }, [result]);
-
-  return result.matches;
-}
